backend/requirements.txt
aiohttp==3.6.2
aniso8601==8.0.0
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-RESTful==0.3.7
gunicorn==20.0.4
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==4.7.3
pylint==2.4.4
pytz==2019.3
six==1.13.0
slackclient==2.5.0
typed-ast==1.4.0
Werkzeug==0.16.0
wrapt==1.11.2
yarl==1.4.2
